Extract ProjectSort
